requirements.txt
absl-py==1.0.0
app==0.0.1
beautifulsoup4==4.11.1
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
easydict==1.9
filelock==3.7.0
flake8==4.0.1
Flask==2.1.1
flask-opencv-streamer==1.4
fonttools==4.33.3
future==0.18.2
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.7.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.3
install==1.3.5
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyGObject==3.42.0
PyGObject-stubs==0.0.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
tb-nightly==2.10.0a20220526
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
